batasan

+++ mv.c
@@ -3,6 +3,10 @@
 #include "user.h"
 
 int main (int argc, char *argv[]){
+  if (argc != 3){
+    printf(2, "Usage: mv oldname newname\n");
+    exit();
+  }
   if ((link(argv[1], argv[2]) < 0) || (unlink(argv[1]) < 0))
     printf(2, "mv %s to %s failed\n", argv[1], argv[2]);
